==> 02-eda.rmd <==
---
title: "02-eda-text"
author: "mcc"
date: "4/16/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Exploratory Data Analysis

>"Exploratory data analysis (EDA) is an approach to analyzing data sets to summarize their main characteristics, often with visual methods." [^21]

[^21]:https://en.wikipedia.org/wiki/Exploratory_data_analysis

## Introduction

The experiment described herein involves taking groups of proteins from the Uniprot.org database and comparing how well different machine learning techniques do at separating the positive from the negative control grouping. In this circumstance, proteins from the myoglobin family are analyzed against randomly chosen human proteins, which are not related to hemoglobin or myoglobin. 

This work is to characterize the *anomalous points* derived from PCA and compare them to the false-positives and false-negatives generated from each of six machine learning approaches produces. For the sake of this paper *anomalous points* are defined as values greater than the absolute value of three times the standard deviation from of the first and second principal components. 

\begin{equation}
Anomalous ~Points > | 3 \sigma | ~~~where~~~ \sigma = \sqrt{ \frac{1}{n-1} \sum^n_{i=1} (x_i - \bar x)^2}
\end{equation}

Therefore the M.L techniques will be: 

1. Principal Component Analysis,
1. Logistic Regression, 
1. SVM-Linear,
1. SVM-polynomial, 
1. SVM-RBF, 
1. Neural Network.

### Four-Step Analysis

At this stage, data is inspected in a careful and structured way. Hence, I have chosen a four-step process:

1. Hypothesize,
1. Summarize,
1. Visualize,
1. Normalize.

### Useful Guides for Exploratory Data Analysis

The summarization of the amino acid dataset is based on a hybrid set of guidelines; 

1. NIST Handbook of Statistics,[^22]
2. Exploratory Data Analysis With R by Roger Peng,[^23]
3. Exploratory Data Analysis Using R by Ronald K. Pearson.[^24] 

[^22]:https://www.itl.nist.gov/div898/handbook/

[^23]:Roger Peng, Exploratory Data Analysis with R, https://leanpub.com/exdata, 2016

[^24]:Ronald Pearson, Exploratory Data Analysis Using R, CRC Press, ISBN:9781138480605, 2018

### Questions During EDA

Although exploratory data analysis does not always have a formal hypothesis testing portion, I do, however, pose several questions concerning the structure, quality, and types of data.

1. Do the independent variables of this study have large skewed distributions?

    1.1 If skews are greater than 2.0, then can a transformation be used for normalization?

    1.2 Determine what transformation to use?

2. Can Feature Selection be used, and which procedures are appropriate?

    2.1 Use the Random Forest technique known as Boruta[^25] for feature importance or reduction?

    2.2 Will coefficients of correlation (R) find collinearity and reduce the number of features?

    2.3 Will principal component analysis (PCA) be useful in finding hidden structures of patterns?

    2.4 Can PCA be used successfully for Feature Selection?

3. What is the structure of the data?

    3.1 Is the data representative of the entire experimental space?

    3.2 Is missing data an issue?

    3.3 Does the data have certain biases, either known or unknown?

    3.4 What relationships do we expect from these variables?[^26]

[^25]:Miron Kursa, Witold Rudnicki, Feature Selection with the Boruta Package, DOI:10.18637/jss.v036.i11, 2010

[^26]:Ronald Pearson, Exploratory Data Analysis Using R, CRC Press, ISBN:9781138480605, 2018

### Conclusion for Boruta random forest test, RAW Data

- All features are essential. None should be dropped.

### Conclusions For EDA, RAW data

Three amino acids (C, F, I) from the single amino acid percent composition were deemed problematic due to their skewness were greater than 2.0. This suggests that a transformation should be carried out to rectify this issue.

| Protein          | Skewness |
| :--------------- | :------: |
| C, Cysteine      | 2.538162 |
| F, Phenolalanine | 2.128118 |
| I, Isoleucine    | 2.192145 |


## Analysis of TRANSFORMED data

**This EDA section is a reevaluation square root transformed, `c_m_RAW_ACC.csv` data set, hence called `c_m_TRANSFORMED.csv.` **

The $\sqrt x_i$ *Transformed* data is derived from `c_m_RAW_ACC.csv` where the amino acids C, F, I were transformed using a square root function. This transformation was done to reduce the skewness of these samples and avoid modeling problems arising from high skewness, as seen below.

| Amino Acid       | Initial skewness | Skew After Square-Root Transformation |
| :--------------- | :--------------: | :-----------------------------------: |
| C, Cysteine      |     2.538162     |               0.3478132               |
| F, Phenolalanine |     2.128118     |               -0.102739               |
| I, Isoleucine    |     2.192145     |               0.2934749               |


## EDA Conclusion

### Feature Selection & Extraction

It was determined early on that three amino acids (C, F, I) from the data amino acid percent compositions (c_m_RAW_AAC.csv) had Skewness greater than two. It was found that tranforming the features using the square root function lowered the skewness to {-0.102739 $\leq$ skew after transformation $\leq$ 0.3478132}.

Table 7.1, Skewness Before And After Square-Root Transform

| Amino Acid       | Initial Skewness | Skew After Square-Root Transform |
| :--------------- | :--------------: | :------------------------------: |
| C, Cysteine      |     2.538162     |           0.347813248            |
| F, Phenolalanine |     2.128118     |          -0.102739748            |
| I, Isoleucine    |     2.192145     |           0.293474879            |

The transformations of the three amino acids (C, F, I) did not appriciably change the Correlation coefficient, R. Therefore no R values were above 0.75 before or after testing. The highest coeffiecient of correlation being Threonine and Argnine with an R of 0.7098. This indicates that no features are collinear. Therefore the transformed data is used throughout this experiment.

---

### Information Block**

How to: Dimension Reduction using High Correlation

How to reduce features given high correlation (|R| >= 0.75) {-}

1. Calculate the correlation matrix of the predictors. 

2. If the correlation plot produced of any two variables is greater than or equal to (|R| >= 0.75), then we could consider feature elimination. This interesting heuristic approach would be used for determining which feature to eliminate.[^29] 

3. Determine if the two predictors associated with the most significant absolute pairwise correlation (R > |0.75|), call them predictors A and B. 

4. Determine the average correlation between A and the other variables. Do the same for predictor B. 

5. If A has a more significant average correlation, remove it; otherwise, remove predictor B. 

6. Repeat Steps 2–4 until no absolute correlations are above the threshold. 

---

An alternative test for variable importance carried out is called Boruta. Boruta builds Random Forests then "finds relevant features by comparing original attributes' importance with importance achievable at random." [^71]

[^71]:https://notabug.org/mbq/Boruta/

Boruta is used for dimensionality reduction of the **c_m_Transformed data**. Bortua showed that all dependent features are essential for the generation of a Random Forest Decision Tree. It would wise to keep all features for that model test and throughout the generation of other models. All features have decisive mean importance, which is generated by a Gini calculation.
