==> 07-results.rmd <==

# Results

I have studied six different M.L. algorithms using protein amino acid percent composition data from two classes. Class number 1 is Myoglobin proteins, which is the positive control set. While the second class is a control group (0) of human proteins that do not have Fe binding centers.

| Group     |  Class   | N of Class | Range of Groups |
| :-------- | :------: | :--------: | --------------: |
| Controls  | 0 or (-) |    1216    |    1, ..., 1216 |
| Myoglobin | 1 or (+) |    1124    | 1217, ..., 2340 |

## The Six M.L Algorithms consist of:

| Name                             |         Type |    Output Used For Graphing |
| :------------------------------- | -----------: | :-------------------------: |
| Principal Component Analysis     | Unsupervised | Anomalies > Abs($3 \sigma$) |
| Logistic Regression              |   Supervised |                     FP & FN |
| SVM-linear                       |   Supervised |                     FP & FN |
| SVM-polynomial kernel            |   Supervised |                     FP & FN |
| SVM-radial basis function kernel |   Supervised |                     FP & FN |
| Neural Network                   |   Supervised |                     FP & FN |

=========================



### Statistical Learning Method Vs Total Number of FP/FN

| Statistical Method                  | Unique |
| :---------------------------------- | -----: |
| Principal Componnent Analysis       |    460 |
| Logit                               |    119 |
| SVM Linear                          |    125 |
| SVM Polynomial                      |     70 |
| SVM Radial Basis Function           |     58 |
| Deep Learning                       |     79 |




### Mean Accuracies of M.L. Techniques, n=10

| Rank |  M.L. Technique | Mean Accuracy |
| ---: | --------------: | ------------: |
|    1 |         SVM-RBF |     0.9510603 |
|    2 |        SVM-Poly |     0.9415091 |
|    3 |         SVM-Lin |     0.9292275 |
|    4 | NN w 20 Neurons |     0.9286350 |
|    5 |           Logit |     0.9078127 |
