==> 09-intro-ds.rmd <==
---
title: "Introduction to Data Science, R and RStudio"
author: "mcc"
date: "3/13/2020"
output: pdf_document
---


# Appendices

```{r}
knitr::opts_chunk$set(cache = TRUE)
```

**Introduction to Data Science, R and RStudio** 

- [LEARN GIT](https://try.github.io/). Get a [Github account!](https://github.com/) 
  - Put Every computer program you write and Every stick of knowledge related to your work in a clean format on Git. You will benefit from it in the long run...

- To Biologists and Biochemists (which I also consider myself), it is your task to become more familiar with computer languages,  computer concepts and math/statistics. 

As of the writing of this booklet, Data Science and Bioinformatics are now centered around the computer languages [R](https://cran.r-project.org/) and [Python](https://www.anaconda.com/).

To many researchers in data science and bioinformatics the field now includes such languages as, in **no particular order**,

- [R](https://cran.r-project.org/)
- [Python](https://www.anaconda.com/)
- [Bash shell scripting](https://www.datascienceatthecommandline.com/)  
- [SQL](https://en.wikipedia.org/wiki/SQL)
- [Javascript](https://www.javascript.com/) 
- [RMarkdown](https://rmarkdown.rstudio.com/), ([This one is fun and easy](https://www.markdowntutorial.com))

#### Just start with one language {-} 

- Stick with it for a time and learn it. Learn the ins-and-outs of one language first.

#### Is R or Python better? {-}

1. R & Python are both FREE,
1. Both have great integrated development environments (IDEs),
   - [RStudio](https://www.rstudio.com/) is great & FREE,
   - [Spyder](https://www.spyder-ide.org/) is also great & FREE,
   - [PyCharm](https://www.jetbrains.com/pycharm/)
   - [Sublime](https://www.sublimetext.com/)
   - [Visual Studio Code](https://code.visualstudio.com/)
   - [Jupyter Notebook](https://jupyter.org/)
1. Both languages have been around for > 20 years, therefore both have tons of FREE information & tutorials on YouTube,

## Install R & RStudio

**NOTE**: I use [Ubuntu](https://www.ubuntu.com/).

1. Go to: https://cran.r-project.org/

1. Choose R for your operating system.

1. If you are using Linux, I recommend that you download/install 4 R files.
   1. r-base-core_#.#.#.*.deb  (approx. 30 MB)
   1. r-base-dev_#.#.#.*.deb   (approx. 45 KB)
   1. r-base_#.#.#*.deb        (approx. 90 KB)
   1. r-base-html-#.#.#*.deb   (approx. 90 KB)

1. If you do have Linux you may try this video: [How to install R](https://www.youtube.com/watch?v=LeB15G14N7g).
   
1. Go to: https://www.rstudio.com/

1. From RStudio's homepage click, *Products* then click *RStudio* from the drop-down menu.

1. Click the Download of the FREE version of RStudio Desktop

1. Click *RStudio #.#* to download a version for your machine

1. Have Linux? Try this - [How to install RStudio.](https://www.youtube.com/watch?v=LeB15G14N7g)

1. If you are looking for instructions for Mac & Windows machines try: 
   - [FreeCodeCamp](https://guide.freecodecamp.org/r/)
   
## Load Libraries Used In This Project
 
If you are using Ubuntu/Linux you may need these Linux libraries first.
```
sudo apt-get install libcurl4-openssl-dev libssl-dev libxml2-dev build-essential
```

To install `car` & `rgl`
```
sudo apt-get install xorg libx11-dev libglu1-mesa-dev libfreetype6-dev
```


```
install.packages("rlang")
library(rlang)

load_or_install <- function(package_names) {
    for(package_name in package_names) {
        if(!is_installed(package_name)) {
            install.packages(package_name,
            repos = "http://lib.stat.cmu.edu/R/CRAN",
            dependencies = TRUE)
        }
        library(package_name,character.only=TRUE,quietly=TRUE,verbose=FALSE) 
        print("OK")
        }
}
```

```
load_or_install(c("doMC", "corrplot", "knitr", "caret", "tidyverse"))

load_or_install(c("ggplot2", "rmarkdown", "bookdown", "blogdown", "kernlab"))

load_or_install(c("e1071", "plyr", "RColorBrewer", "neuralnet", "ggfortify"))

load_or_install(c("rpart", "MASS", "tidyr", "ggplot2", "seqinr", "Boruta", "kableExtra"))

load_or_install(c("LogicReg", "randomForest", "foreach", "caretEnsemble"))

load_or_install(c("import", "dplyr", "stringr", "stringi", "readr", "tinytex"))
```



## Where To Find Help

1. [Cheat Sheets](https://resources.rstudio.com/rstudio-cheatsheets)
1. https://community.rstudio.com
1. https://www.reddit.com/r/RStudio/
1. https://R-bloggers.com/
1. https://resources.rstudio.com/
1. [Rpubs.com](https://rpubs.com/)

- **Rpubs.com** contains R/RStudio notebooks and Markdown pages, VERY HELPFUL work from other peoples online R documents. It is a way to learn from others and share your work. - Sign up, it is [FREE](https://rpubs.com/)! then press: *Get Started*

**NOTE**: If you are interested in seeing what others have published search Google, Rpubs.com does not have its own search function. In Google, Search: `site:rpubs.com eda`

Other sites:

1. [Coursera](https://www.coursera.org/)
1. [Stack Overflow](https://stackoverflow.com/)
1. [Quora](https://www.quora.com/What-is-data-science)
1. [Roger Peng's EDA](https://youtu.be/ZXrHkIz-krE?t=15)
1. [Bookdown](https://bookdown.org/) - **terrible** yet necessary resource

The Lean Publishing (https://leanpub.com) company contains a library in the form of FREE down-loadable books/pdfs. I recommend;

1. How to be a modern scientist[^91]  by Jeffrey Leek[^92] 
2. R Programming for Data Science[^93]  by Roger Peng[^94] 
3. Exploratory Data Analysis with R[^95]  by Roger Peng
4. Data Analysis for the Life Sciences[^96]  by Rafael Irizarry & Michael Love

[^91]:https://leanpub.com/modernscientist
[^92]:http://jtleek.com
[^93]:https://leanpub.com/rprogramming
[^94]:https://simplystatistics.org
[^95]:https://leanpub.com/exdata
[^96]:https://leanpub.com/dataanalysisforthelifesciences

## Machine Setting & Session Info

```{r, include=FALSE}
library(knitr)
opts_chunk$set(cache = TRUE, warning = FALSE, message = FALSE)
```

```{r}
Sys.info()[c(1:3,5)]
```

```{r}
sessionInfo()
```
